refactor: extract comment CRUD logic into composable and shared utilities
