+++ app/utils/avatar.ts
@@ -1,0 +1,34 @@
+const avatarPalette = [
+  'bg-violet-100 text-violet-700 dark:bg-violet-900/40 dark:text-violet-300',
+  'bg-sky-100 text-sky-700 dark:bg-sky-900/40 dark:text-sky-300',
+  'bg-amber-100 text-amber-700 dark:bg-amber-900/40 dark:text-amber-300',
+  'bg-emerald-100 text-emerald-700 dark:bg-emerald-900/40 dark:text-emerald-300',
+  'bg-rose-100 text-rose-700 dark:bg-rose-900/40 dark:text-rose-300',
+  'bg-indigo-100 text-indigo-700 dark:bg-indigo-900/40 dark:text-indigo-300',
+  'bg-teal-100 text-teal-700 dark:bg-teal-900/40 dark:text-teal-300',
+  'bg-orange-100 text-orange-700 dark:bg-orange-900/40 dark:text-orange-300'
+]
+
+/**
+ * Derive 2-character initials from a name or email string.
+ */
+export function getInitials(nameOrEmail: string | null): string {
+  if (!nameOrEmail) return '?'
+  const parts = nameOrEmail.trim().split(/[\s@.]+/).filter(Boolean)
+  if (parts.length >= 2) {
+    return (parts[0]!.charAt(0) + parts[1]!.charAt(0)).toUpperCase()
+  }
+  return nameOrEmail.slice(0, 2).toUpperCase()
+}
+
+/**
+ * Deterministic avatar background color from any unique ID string.
+ */
+export function getAvatarColor(id: string): string {
+  let hash = 0
+  for (let i = 0; i < id.length; i++) {
+    hash = ((hash << 5) - hash) + id.charCodeAt(i)
+    hash |= 0
+  }
+  return avatarPalette[Math.abs(hash) % avatarPalette.length]!
+}
