+++ server/api/projects/[projectId]/issues/[issueId]/comments.get.ts
@@ -1,7 +1,8 @@
 import { eq, asc } from 'drizzle-orm'
-import { issueComments, issues, profiles } from '~~/server/database/schema'
-import { badRequest, notFound } from '~~/server/utils/errors'
+import { issueComments, profiles } from '~~/server/database/schema'
+import { badRequest } from '~~/server/utils/errors'
 import { getAccessibleProject } from '~~/server/utils/project-access'
+import { getProjectIssue } from '~~/server/utils/comment-access'
 
 export default defineEventHandler(async (event) => {
   const projectId = getRouterParam(event, 'projectId')
@@ -15,17 +16,7 @@ export default defineEventHandler(async (event) => {
   const userId = event.context.userId as string
 
   await getAccessibleProject(db, projectId, userId)
-
-  // Verify issue exists and belongs to project
-  const [issue] = await db
-    .select({ id: issues.id })
-    .from(issues)
-    .where(eq(issues.id, Number(issueId)))
-    .limit(1)
-
-  if (!issue) {
-    notFound('Issue not found')
-  }
+  await getProjectIssue(db, projectId, issueId)
 
   const comments = await db
     .select({
