+++ server/api/projects/[projectId]/issues/[issueId]/comments.post.ts
@@ -1,9 +1,10 @@
 import { eq } from 'drizzle-orm'
-import { issueComments, issues, notifications, profiles } from '~~/server/database/schema'
+import { issueComments, notifications, profiles } from '~~/server/database/schema'
 import { createCommentSchema } from '~~/shared/schemas/issue-comment'
 import { NotificationType } from '~~/shared/constants'
-import { badRequest, notFound } from '~~/server/utils/errors'
+import { badRequest } from '~~/server/utils/errors'
 import { getAccessibleProject } from '~~/server/utils/project-access'
+import { getProjectIssue } from '~~/server/utils/comment-access'
 import { sanitizeHtml, stripHtmlTags } from '~~/server/utils/html'
 
 export default defineEventHandler(async (event) => {
@@ -18,23 +19,7 @@ export default defineEventHandler(async (event) => {
   const userId = event.context.userId as string
 
   await getAccessibleProject(db, projectId, userId)
-
-  // Verify issue exists and belongs to project
-  const [issue] = await db
-    .select({
-      id: issues.id,
-      title: issues.title,
-      projectKey: issues.projectKey,
-      issueNumber: issues.issueNumber,
-      createdBy: issues.createdBy
-    })
-    .from(issues)
-    .where(eq(issues.id, Number(issueId)))
-    .limit(1)
-
-  if (!issue) {
-    notFound('Issue not found')
-  }
+  const issue = await getProjectIssue(db, projectId, issueId)
 
   const body = await readBody(event)
   const result = createCommentSchema.safeParse(body)
