+++ server/api/projects/[projectId]/issues/[issueId]/comments/[commentId].delete.ts
@@ -1,7 +1,8 @@
-import { and, eq } from 'drizzle-orm'
-import { issueComments, issues } from '~~/server/database/schema'
-import { badRequest, forbidden, notFound } from '~~/server/utils/errors'
+import { eq } from 'drizzle-orm'
+import { issueComments } from '~~/server/database/schema'
+import { badRequest } from '~~/server/utils/errors'
 import { getAccessibleProject } from '~~/server/utils/project-access'
+import { getProjectIssue, getIssueComment, assertCommentAuthor } from '~~/server/utils/comment-access'
 
 export default defineEventHandler(async (event) => {
   const projectId = getRouterParam(event, 'projectId')
@@ -16,45 +17,13 @@ export default defineEventHandler(async (event) => {
   const userId = event.context.userId as string
 
   await getAccessibleProject(db, projectId, userId)
-
-  // Verify issue belongs to this project
-  const [issue] = await db
-    .select({ id: issues.id })
-    .from(issues)
-    .where(
-      and(
-        eq(issues.id, Number(issueId)),
-        eq(issues.projectId, projectId)
-      )
-    )
-    .limit(1)
-
-  if (!issue) {
-    notFound('Issue not found')
-  }
-
-  const [comment] = await db
-    .select({ id: issueComments.id, authorId: issueComments.authorId })
-    .from(issueComments)
-    .where(
-      and(
-        eq(issueComments.id, Number(commentId)),
-        eq(issueComments.issueId, Number(issueId))
-      )
-    )
-    .limit(1)
-
-  if (!comment) {
-    notFound('Comment not found')
-  }
-
-  if (comment.authorId !== userId) {
-    forbidden('You can only delete your own comments')
-  }
+  await getProjectIssue(db, projectId, issueId)
+  const comment = await getIssueComment(db, issueId, commentId)
+  assertCommentAuthor(comment, userId, 'delete')
 
   await db
     .delete(issueComments)
-    .where(eq(issueComments.id, Number(commentId)))
+    .where(eq(issueComments.id, comment.id))
 
   return { success: true }
 })
