+++ server/api/projects/[projectId]/issues/[issueId]/comments/[commentId].patch.ts
@@ -1,8 +1,9 @@
-import { and, eq } from 'drizzle-orm'
-import { issueComments, issues } from '~~/server/database/schema'
+import { eq } from 'drizzle-orm'
+import { issueComments } from '~~/server/database/schema'
 import { updateCommentSchema } from '~~/shared/schemas/issue-comment'
-import { badRequest, forbidden, notFound } from '~~/server/utils/errors'
+import { badRequest } from '~~/server/utils/errors'
 import { getAccessibleProject } from '~~/server/utils/project-access'
+import { getProjectIssue, getIssueComment, assertCommentAuthor } from '~~/server/utils/comment-access'
 import { sanitizeHtml } from '~~/server/utils/html'
 
 export default defineEventHandler(async (event) => {
@@ -18,41 +19,9 @@ export default defineEventHandler(async (event) => {
   const userId = event.context.userId as string
 
   await getAccessibleProject(db, projectId, userId)
-
-  // Verify issue belongs to this project
-  const [issue] = await db
-    .select({ id: issues.id })
-    .from(issues)
-    .where(
-      and(
-        eq(issues.id, Number(issueId)),
-        eq(issues.projectId, projectId)
-      )
-    )
-    .limit(1)
-
-  if (!issue) {
-    notFound('Issue not found')
-  }
-
-  const [comment] = await db
-    .select({ id: issueComments.id, authorId: issueComments.authorId })
-    .from(issueComments)
-    .where(
-      and(
-        eq(issueComments.id, Number(commentId)),
-        eq(issueComments.issueId, Number(issueId))
-      )
-    )
-    .limit(1)
-
-  if (!comment) {
-    notFound('Comment not found')
-  }
-
-  if (comment.authorId !== userId) {
-    forbidden('You can only edit your own comments')
-  }
+  await getProjectIssue(db, projectId, issueId)
+  const comment = await getIssueComment(db, issueId, commentId)
+  assertCommentAuthor(comment, userId, 'edit')
 
   const body = await readBody(event)
   const result = updateCommentSchema.safeParse(body)
@@ -69,7 +38,7 @@ export default defineEventHandler(async (event) => {
       content: sanitizedContent,
       updatedAt: new Date()
     })
-    .where(eq(issueComments.id, Number(commentId)))
+    .where(eq(issueComments.id, comment.id))
     .returning()
 
   const profile = event.context.profile as { name: string | null, email: string }
